// src/services/types.ts
export type Character = {
    id: number;
    created: string;
    episode?: string[];
    gender: string;
    image: string;
    location: {
      name: string;
      url?: string;
    };
    name: string;
    origin?: {
      name: string;
      url: string;
    };
    species: string;
    status: string;
    type: string;
    url?: string;
  };

export type AllCharactersResponse = {
    info: {
      count: number;
      next: string | null;
      pages: number;
      prev: string | null;
    };
    results: Character[];
  };